refactor: divide a test file into 2
